+++ functions.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Avidly Theme functions and definitions.
+ * Avidly Block Theme functions and definitions.
  *
  * Include PHP files, nothing else.
  *
@@ -15,33 +15,32 @@
 /**
  * Configuration
  */
+require_once 'inc/_conf/wp-cleanup.php'; // Remove unneeded WordPress default functionality.
 require_once 'inc/_conf/cache-busting.php'; // Purge assets cache in register-assets.php.
 require_once 'inc/_conf/setup-theme.php';
 require_once 'inc/_conf/setup-editor.php';
 require_once 'inc/_conf/register-assets.php';
-require_once 'inc/_conf/register-menus.php';
-require_once 'inc/_conf/register-sidebars.php';
+// require_once 'inc/_conf/register-menus.php'; // Uncomment his if you wan't to use Classic menus. You need also a custom block or Avidly Classic Menu block to display classic menus from backend.
 require_once 'inc/_conf/register-block-patterns.php';
 require_once 'inc/_conf/register-block-templates.php';
-
-/**
- * Classes
- */
-require_once 'inc/class/class-disclosure-nav-menu.php';
-require_once 'inc/class/class-svg-icons.php';
+require_once 'inc/_conf/register-block-styles.php';
 
 /**
  * Helpers
  */
-require_once 'inc/helpers/wp-cleanup.php'; // Remove unneeded WordPress default functionality.
-require_once 'inc/helpers/svg-icons.php';
-require_once 'inc/helpers/featured-images.php';
-require_once 'inc/helpers/override-core-styles.php';
 require_once 'inc/helpers/theme-hooks.php';
+require_once 'inc/helpers/override-core-styles.php';
+require_once 'inc/helpers/featured-images.php';
 require_once 'inc/helpers/helper-functions.php';
+
+/**
+ * Block queries
+ */
+require_once 'inc/queries/related-posts-query.php';
 
 /**
  * Block renders
  */
-require_once 'inc/render-block/post-template.php';
 require_once 'inc/render-block/image.php';
+require_once 'inc/render-block/post-template.php';
+require_once 'inc/render-block/post-terms.php';