+++ inc/_conf/setup-editor.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Setup editor properties and editor related functionality
+ * Setup editor properties.
  *
  * @link https://developer.wordpress.org/block-editor/how-to-guides/themes/theme-support/
  *
@@ -8,38 +8,25 @@
  */
 
 // phpcs:disable
-add_filter( 'allowed_block_types_all', 'avidly_theme_set_default_blocks', 10, 2 );
+add_filter( 'allowed_block_types_all', 'avidly_theme_set_core_blocks', 10, 2 );
+add_filter( 'allowed_block_types_all', 'avidly_theme_set_extended_blocks', 15, 2 );
 // add_filter( 'allowed_block_types_all', 'avidly_theme_set_wc_blocks', 15, 2 );
 // add_filter( 'allowed_block_types_all', 'avidly_theme_set_ld_blocks', 15, 2 );
 // add_filter( 'allowed_block_types_all', 'avidly_theme_set_cap_based_blocks', 15, 2 );
 // phpcs:enable
-add_filter( 'block_editor_settings_all', 'avidly_theme_default_block_template' );
-
-
-/**
- * Filters the settings to pass to the block editor for all editor type.
- *
- * @param array $settings Default editor settings.
- * @return $settings
- */
-function avidly_theme_default_block_template( $settings ) {
-	$settings['defaultBlockTemplate'] = file_get_contents( get_theme_file_path( 'templates/block-template.html' ) );
-	return $settings;
-}
-
 
 /**
  * Whitelist core blocks
  *
  * Comment out the ones theme doesn't use.
  *
- * @param bool|string             $allowed_block_types Array of block type slugs, or boolean to enable/disable all.
+ * @param bool|string $allowed_block_types Array of block type slugs, or boolean to enable/disable all.
  * @param WP_Block_Editor_Context $editor_context The current block editor context.
  *
  * @link https://wordpress.org/support/article/blocks/
  * @link https://developer.wordpress.org/reference/hooks/allowed_block_types_all/
  */
-function avidly_theme_set_default_blocks( $allowed_block_types, $editor_context ) {
+function avidly_theme_set_core_blocks( $allowed_block_types, $editor_context ) {
 
 	// DO not use whitelist if block editor context is not detected.
 	if ( empty( $editor_context->post ) ) {
@@ -67,7 +54,7 @@ function avidly_theme_set_default_blocks( $allowed_block_types, $editor_context 
 		// 'core/freeform', // Classic editor.
 		'core/html',
 		// 'core/preformatted',
-		// 'core/pullquote',
+		'core/pullquote',
 		'core/table',
 		// 'core/verse',
 
@@ -83,7 +70,7 @@ function avidly_theme_set_default_blocks( $allowed_block_types, $editor_context 
 		'core/media-text',
 		// 'core/more',
 		'core/block', // Reusable block.
-		// 'core/separator',
+		'core/separator',
 		'core/missing', // New in core 6.0.
 
 		// Widgets.
@@ -92,14 +79,14 @@ function avidly_theme_set_default_blocks( $allowed_block_types, $editor_context 
 		// 'core/calendar', // New in core 5.3.
 		'core/categories',
 		// 'core/latest-comments',
-		'core/latest-posts',
-		'core/rss', // New in core 5.3.
-		// 'core/social-link', // New in core 5.4, required for social-links.
-		// 'core/social-links', // New in core 5.4.
-		// 'core/search', // New in core 5.4.
-		'core/tag-cloud', // New in core 5.3.
-		'core/widget-group', // New in core 5.8.
-		'core/legacy-widget', // New in core 5.8.
+		// 'core/latest-posts',
+		// 'core/rss', // New in core 5.3.
+		'core/social-link', // New in core 5.4, required for social-links.
+		'core/social-links', // New in core 5.4.
+		'core/search', // New in core 5.4.
+		// 'core/tag-cloud', // New in core 5.3.
+		// 'core/widget-group', // New in core 5.8.
+		// 'core/legacy-widget', // New in core 5.8.
 
 		// Theme blocks.
 		'core/site-logo', // New in core 5.8.
@@ -113,6 +100,7 @@ function avidly_theme_set_default_blocks( $allowed_block_types, $editor_context 
 		'core/query-pagination-next', // New in core 5.8.
 		'core/query-pagination-previous', // New in core 5.8.
 		'core/query-no-results',  // New in core 6.0.
+		'core/template-part', // New in core 5.9.
 		'core/term-description', // New in core 5.9.
 		'core/post-title', // New in core 5.8.
 		'core/post-content', // New in core 5.8.
@@ -126,16 +114,17 @@ function avidly_theme_set_default_blocks( $allowed_block_types, $editor_context 
 		'core/post-navigation-link', // New in core 6.0.
 		'core/loginout', // New in core 5.8.
 		'core/pagelist', // New in core 5.8.
-		'core/page-list', // Same as above, just renamed in core 6.0?
+		'core/page-list', // New in core 6.0
+		'core/page-list-item', // New in core 6.0, required for page-list.
 		'core/navigation', // New in core 5.9.
 		'core/navigation-link', // New in core 5.9.
 		'core/navigation-submenu', // New in core 5.9.
-		'core/template-part', // New in core 5.9.
 		'core/pattern', // New in core 6.0.
 
 		// Author.
 		'core/avatar', // New in core 6.0.
 		'core/post-author', // New in core 5.9.
+		'core/post-author-name', // New in core 6.0.
 		'core/post-author-biography', // New in core 6.0.
 
 		// Comments, all new in core 6.0.
@@ -157,24 +146,7 @@ function avidly_theme_set_default_blocks( $allowed_block_types, $editor_context 
 		// Embed blocks was refactored to be variations of the base Embed block on core 5.6.
 		// Variations can be set via /assets/js/editor-script-block.js.
 		'core/embed',
-
-		// Gravity Forms.
-		'gravityforms/form',
-
-		// Yoast SEO.
-		'yoast-seo/breadcrumbs',
-		// 'yoast/how-to-block',
-		// 'yoast/faq-block',
-
-		// Advanced Custom Fields.
-		// 'acf/your-block-name',
-
-		// Avidly blocks.
-		'avidly/social-share',
-		'avidly/accordion',
-		'avidly/block-toc',
-		'avidly/block-query',
-
+		
 		// phpcs:enable
 	);
 
@@ -182,43 +154,142 @@ function avidly_theme_set_default_blocks( $allowed_block_types, $editor_context 
 }
 
 /**
+ * Enable Misc blocks (example from plugins).
+ *
+ * @param bool|string $allowed_block_types Array of block type slugs, or boolean to enable/disable all.
+ * @param WP_Block_Editor_Context $editor_context The current block editor context.
+ */
+function avidly_theme_set_extended_blocks( $allowed_block_types, $editor_context ) {
+
+	// phpcs:disable
+	$push_blocks = array(
+		// Avidly.
+		'avidly/accordion',
+		'avidly/block-toc',
+		'avidly/social-share',
+
+		// Yoast SEO.
+		'yoast-seo/breadcrumbs',
+		// 'yoast/how-to-block',
+		// 'yoast/faq-block',
+
+		// Gravity forms.
+		'gravityforms/form',
+
+		// Carousel Slider Block for Gutenberg.
+		'cb/carousel', 
+		'cb/slide',
+
+		// Advanced Custom Fields.
+		// 'acf/your-block-name',
+
+	);
+	// phpcs:enable
+
+	$allowed_block_types = array_merge( $allowed_block_types, $push_blocks );
+
+	return $allowed_block_types;
+}
+
+/**
  * Enable WooCommerce blocks.
  *
- * @param bool|string             $allowed_block_types Array of block type slugs, or boolean to enable/disable all.
+ * @param bool|string $allowed_block_types Array of block type slugs, or boolean to enable/disable all.
  * @param WP_Block_Editor_Context $editor_context The current block editor context.
  */
 function avidly_theme_set_wc_blocks( $allowed_block_types, $editor_context ) {
 
-	// Updated list from WC version 6.5.1.
+	// Updated list from WC version 7.7.0.
 	$push_blocks = array(
-		'woocommerce/all-reviews',
-		'woocommerce/featured-category',
-		'woocommerce/featured-product',
-		'woocommerce/handpicked-products',
-		'woocommerce/product-best-sellers',
-		'woocommerce/product-categories',
-		'woocommerce/product-category',
-		'woocommerce/product-new',
-		'woocommerce/product-on-sale',
-		'woocommerce/products-by-attribute',
-		'woocommerce/product-top-rated',
-		'woocommerce/reviews-by-product',
-		'woocommerce/reviews-by-category',
-		'woocommerce/product-search',
-		'woocommerce/product-tag',
-		'woocommerce/product-title',
-		'woocommerce/product-price',
-		'woocommerce/product-image',
-		'woocommerce/product-rating',
-		'woocommerce/product-button',
-		'woocommerce/product-summary',
-		'woocommerce/product-sale-badge',
-		'woocommerce/all-products',
-		'woocommerce/price-filter',
-		'woocommerce/attribute-filter',
-		'woocommerce/stock-filter',
-		'woocommerce/active-filters',
-		'woocommerce/legacy-template',
+		'woocommerce/active-filters', 
+		'woocommerce/product-title', 
+		'woocommerce/product-price', 
+		'woocommerce/product-image', 
+		'woocommerce/product-rating', 
+		'woocommerce/product-button', 
+		'woocommerce/product-summary', 
+		'woocommerce/product-sale-badge', 
+		'woocommerce/product-sku', 
+		'woocommerce/product-stock-indicator', 
+		'woocommerce/all-products', 
+		'woocommerce/all-reviews', 
+		'woocommerce/attribute-filter', 
+		'woocommerce/customer-account', 
+		'woocommerce/featured-category', 
+		'woocommerce/featured-product', 
+		'woocommerce/filter-wrapper', 
+		'woocommerce/handpicked-products', 
+		'woocommerce/mini-cart', 
+		'woocommerce/price-filter', 
+		'woocommerce/product-best-sellers', 
+		'woocommerce/product-categories', 
+		'woocommerce/product-category', 
+		'woocommerce/product-new', 
+		'woocommerce/product-on-sale', 
+		'woocommerce/product-search', 
+		'woocommerce/product-tag', 
+		'woocommerce/product-top-rated', 
+		'woocommerce/products-by-attribute', 
+		'woocommerce/rating-filter', 
+		'woocommerce/reviews-by-category', 
+		'woocommerce/reviews-by-product', 
+		'woocommerce/stock-filter', 
+		'woocommerce/filled-cart-block', 
+		'woocommerce/cart-items-block', 
+		'woocommerce/cart-line-items-block', 
+		'woocommerce/cart-cross-sells-block', 
+		'woocommerce/cart-cross-sells-products-block', 
+		'woocommerce/cart-totals-block', 
+		'woocommerce/cart-express-payment-block', 
+		'woocommerce/proceed-to-checkout-block', 
+		'woocommerce/empty-cart-block', 
+		'woocommerce/cart-accepted-payment-methods-block', 
+		'woocommerce/cart-order-summary-block', 
+		'woocommerce/cart-order-summary-subtotal-block', 
+		'woocommerce/cart-order-summary-fee-block', 
+		'woocommerce/cart-order-summary-discount-block', 
+		'woocommerce/cart-order-summary-shipping-block', 
+		'woocommerce/cart-order-summary-coupon-form-block', 
+		'woocommerce/cart-order-summary-taxes-block', 
+		'woocommerce/cart-order-summary-heading-block', 
+		'woocommerce/cart', 
+		'woocommerce/checkout-fields-block', 
+		'woocommerce/checkout-totals-block', 
+		'woocommerce/checkout-shipping-address-block', 
+		'woocommerce/checkout-terms-block', 
+		'woocommerce/checkout-contact-information-block', 
+		'woocommerce/checkout-billing-address-block', 
+		'woocommerce/checkout-actions-block', 
+		'woocommerce/checkout-order-note-block', 
+		'woocommerce/checkout-order-summary-block', 
+		'woocommerce/checkout-payment-block', 
+		'woocommerce/checkout-express-payment-block', 
+		'woocommerce/checkout-shipping-method-block', 
+		'woocommerce/checkout-shipping-methods-block', 
+		'woocommerce/checkout-pickup-options-block', 
+		'woocommerce/checkout-order-summary-subtotal-block', 
+		'woocommerce/checkout-order-summary-fee-block', 
+		'woocommerce/checkout-order-summary-discount-block', 
+		'woocommerce/checkout-order-summary-shipping-block', 
+		'woocommerce/checkout-order-summary-coupon-form-block', 
+		'woocommerce/checkout-order-summary-taxes-block', 
+		'woocommerce/checkout-order-summary-cart-items-block', 
+		'woocommerce/checkout', 
+		'woocommerce/empty-mini-cart-contents-block', 
+		'woocommerce/filled-mini-cart-contents-block', 
+		'woocommerce/mini-cart-title-block', 
+		'woocommerce/mini-cart-items-block', 
+		'woocommerce/mini-cart-products-table-block', 
+		'woocommerce/mini-cart-footer-block', 
+		'woocommerce/mini-cart-shopping-button-block', 
+		'woocommerce/mini-cart-cart-button-block', 
+		'woocommerce/mini-cart-checkout-button-block', 
+		'woocommerce/mini-cart-contents',
+		'woocommerce/add-to-cart-form', 
+		'woocommerce/product-image-gallery', 
+		'woocommerce/product-details', 
+		'woocommerce/product-reviews', 
+		'woocommerce/product-meta'
 	);
 
 	$allowed_block_types = array_merge( $allowed_block_types, $push_blocks );
@@ -229,50 +300,50 @@ function avidly_theme_set_wc_blocks( $allowed_block_types, $editor_context ) {
 /**
  * Example: Enable LearnDash blocks.
  *
- * @param bool|string             $allowed_block_types Array of block type slugs, or boolean to enable/disable all.
+ * @param bool|string $allowed_block_types Array of block type slugs, or boolean to enable/disable all.
  * @param WP_Block_Editor_Context $editor_context The current block editor context.
  */
 function avidly_theme_set_ld_blocks( $allowed_block_types, $editor_context ) {
 
 	// Updated list from LD version 4.2.0.1.
 	$push_blocks = array(
-		'learndash/ld-login',
-		'learndash/ld-profile',
-		'learndash/ld-course-list',
-		'learndash/ld-lesson-list',
-		'learndash/ld-topic-list',
-		'learndash/ld-quiz-list',
-		'learndash/ld-course-progress',
-		'learndash/ld-visitor',
-		'learndash/ld-student',
-		'learndash/ld-course-complete',
-		'learndash/ld-course-inprogress',
-		'learndash/ld-course-notstarted',
-		'learndash/ld-course-resume',
-		'learndash/ld-course-info',
-		'learndash/ld-user-course-points',
-		'learndash/ld-group-list',
-		'learndash/ld-user-groups',
-		'learndash/ld-group',
-		'learndash/ld-payment-buttons',
-		'learndash/ld-course-content',
-		'learndash/ld-course-expire-status',
-		'learndash/ld-certificate',
-		'learndash/ld-quiz-complete',
-		'learndash/ld-courseinfo',
-		'learndash/ld-quizinfo',
-		'learndash/ld-groupinfo',
-		'learndash/ld-usermeta',
-		'learndash/ld-registration',
-		'learndash/ld-infobar',
-		'learndash/ld-materials',
-		'learndash/ld-user-status',
-		'learndash/ld-navigation',
-		'learndash/ld-exam-question',
-		'learndash/ld-question-answers-block',
-		'learndash/ld-incorrect-answer-message-block',
-		'learndash/ld-correct-answer-message-block',
-		'learndash/ld-question-description',
+		'learndash/ld-login', 
+		'learndash/ld-profile', 
+		'learndash/ld-course-list', 
+		'learndash/ld-lesson-list', 
+		'learndash/ld-topic-list', 
+		'learndash/ld-quiz-list', 
+		'learndash/ld-course-progress', 
+		'learndash/ld-visitor', 
+		'learndash/ld-student', 
+		'learndash/ld-course-complete', 
+		'learndash/ld-course-inprogress', 
+		'learndash/ld-course-notstarted', 
+		'learndash/ld-course-resume', 
+		'learndash/ld-course-info', 
+		'learndash/ld-user-course-points', 
+		'learndash/ld-group-list', 
+		'learndash/ld-user-groups', 
+		'learndash/ld-group', 
+		'learndash/ld-payment-buttons', 
+		'learndash/ld-course-content', 
+		'learndash/ld-course-expire-status', 
+		'learndash/ld-certificate', 
+		'learndash/ld-quiz-complete', 
+		'learndash/ld-courseinfo', 
+		'learndash/ld-quizinfo', 
+		'learndash/ld-groupinfo', 
+		'learndash/ld-usermeta', 
+		'learndash/ld-registration', 
+		'learndash/ld-infobar', 
+		'learndash/ld-materials', 
+		'learndash/ld-user-status', 
+		'learndash/ld-navigation', 
+		'learndash/ld-exam-question', 
+		'learndash/ld-question-answers-block', 
+		'learndash/ld-incorrect-answer-message-block', 
+		'learndash/ld-correct-answer-message-block', 
+		'learndash/ld-question-description', 
 	);
 
 	$allowed_block_types = array_merge( $allowed_block_types, $push_blocks );
@@ -284,7 +355,7 @@ function avidly_theme_set_ld_blocks( $allowed_block_types, $editor_context ) {
 /**
  * Example: Modify allowed blocks by current user capability.
  *
- * @param bool|string             $allowed_block_types Array of block type slugs, or boolean to enable/disable all.
+ * @param bool|string $allowed_block_types Array of block type slugs, or boolean to enable/disable all.
  * @param WP_Block_Editor_Context $editor_context The current block editor context.
  *
  * @link https://developer.wordpress.org/reference/functions/wp_get_current_user/.
@@ -297,13 +368,13 @@ function avidly_theme_set_cap_based_blocks( $allowed_block_types, $editor_contex
 	// SELECT JUST ONE WAY OF WORK FROM BELOW (either work with unset OR set own whitelists for specific capabilities).
 	// Capability based whitelisting is not tested in production yeat.
 
-	// Way 1: Unallow these blocks for users who cannot edit theme settings (since 2.1.1.).
+	// WAY 1: Unallow these blocks for users who cannot edit theme settings (since 2.1.1.).
 	if ( ! array_key_exists( 'edit_theme_options', $caps ) ) {
 		array_splice( $allowed_block_types, array_search( 'core/html', $allowed_block_types, true ), 1 );
 	}
 	$allowed_block_types = array_values( $allowed_block_types ); // Rebase array keys.
 
-	// Way 2: Allow these blocks for users who cannot edit theme settings (since 2.1.1.).
+	// WAY 2: Allow these blocks for users who cannot edit theme settings (since 2.1.1.).
 	if ( ! array_key_exists( 'edit_theme_options', $caps ) ) {
 		$allowed_block_types = array(
 			'core/paragraph',